Disease, NavigationView, PlantData
Added Disease API Struct.swift but not sending that to Core Data.

I commented out Disease.swift because I realized we don't need it, I will remove later, I need to catch up with main.

DiseaseApiResultDetails, Item and List are added.

PlantDiseases acts as ExploreApi but an interface for Diseases.